+++ examples/storybook/stories/Plugins/Buttons/ButtonsPluginStory.js
@@ -6,8 +6,6 @@ import {
   Page,
 } from '../../Components/StoryParts';
 
-import { convertFromRaw, createWithContent } from 'wix-rich-content-editor';
-
 import ButtonsContentState from '../../../../../e2e/tests/fixtures/buttons.json';
 import ButtonsEditor from './ButtonsEditor';
 import editorSourcecode from '!!raw-loader!./ButtonsEditor.js';
@@ -16,7 +14,6 @@ import viewerSourcecode from '!!raw-loader!./ButtonsViewer.js';
 import TabsWrapper from '../../Components/TabsWrapper';
 import apiData from '../apiData';
 
-const editorState = createWithContent(convertFromRaw(ButtonsContentState));
 export default () => {
   return (
     <TabsWrapper apiData={apiData.BUTTON}>
@@ -28,11 +25,11 @@ export default () => {
           triggers the `onClick` action
         </h4>
         <Section type={Section.Types.COMPARISON}>
-          <RichContentEditorBox sourcecode={editorSourcecode} contentState={ButtonsContentState}>
-            <ButtonsEditor editorState={editorState} />
+          <RichContentEditorBox sourcecode={editorSourcecode} content={ButtonsContentState}>
+            <ButtonsEditor content={ButtonsContentState} />
           </RichContentEditorBox>
           <RichContentViewerBox sourcecode={viewerSourcecode}>
-            <ButtonsViewer initialState={ButtonsContentState} />
+            <ButtonsViewer content={ButtonsContentState} />
           </RichContentViewerBox>
         </Section>
       </Page>