+++ examples/storybook/stories/Plugins/LinkPreview/LinkPreviewEditor.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import { RichContentWrapper } from 'wix-rich-content-wrapper';
-import { RichContentEditor } from 'wix-rich-content-editor';
+import { RicosEditor } from 'ricos-editor';
 import { pluginLinkPreview, LinkPreviewProviders } from 'wix-rich-content-plugin-link-preview';
 import { pluginLink } from 'wix-rich-content-plugin-link';
 import { pluginHtml } from 'wix-rich-content-plugin-html';
@@ -20,14 +19,10 @@ const plugins = [
   pluginHtml(),
 ];
 
-const LinkPreviewEditor = ({ editorState }) => (
-  <RichContentWrapper plugins={plugins} isEditor>
-    <RichContentEditor editorState={editorState} />
-  </RichContentWrapper>
-);
+const LinkPreviewEditor = ({ content }) => <RicosEditor plugins={plugins} content={content} />;
 
 LinkPreviewEditor.propTypes = {
-  editorState: PropTypes.object,
+  content: PropTypes.object,
 };
 
 export default LinkPreviewEditor;