+++ examples/storybook/stories/Plugins/Video/VideoViewer.js
@@ -1,14 +1,12 @@
 import React from 'react';
-import { RichContentViewer } from 'wix-rich-content-viewer';
-import { videoTypeMapper } from 'wix-rich-content-plugin-video/dist/module.viewer';
+import { RicosViewer } from 'ricos-viewer';
+import { pluginVideo } from 'wix-rich-content-plugin-video/dist/module.viewer';
 import PropTypes from 'prop-types';
 
-const Viewer = ({ initialState }) => (
-  <RichContentViewer initialState={initialState} typeMappers={[videoTypeMapper]} />
-);
+const Viewer = ({ content }) => <RicosViewer content={content} plugins={[pluginVideo()]} />;
 
 Viewer.propTypes = {
-  initialState: PropTypes.object,
+  content: PropTypes.object,
 };
 
 export default Viewer;