+++ packages/editor-common/web/src/Base/createBasePlugin.jsx
@@ -3,9 +3,11 @@ import createBaseComponent from './createBaseComponent';
 import createAtomicPluginToolbar from './toolbars/createAtomicPluginToolbar';
 import createInlinePluginToolbar from './toolbars/createInlinePluginToolbar';
 import createInsertPluginButton from './createBaseInsertPluginButton';
+import { generateInsertPluginButtonProps } from '../Utils/generateInsertPluginButtonProps';
 import { deleteBlock, setEntityData } from '../Utils/draftUtils';
 import { simplePubsub } from '../Utils/simplePubsub';
 import { getToolbarTheme } from '../Utils/getToolbarTheme';
+import { TOOLBARS } from '../consts';
 
 const getData = (contentBlock, { getEditorState }) => () =>
   getEditorState()
@@ -40,11 +42,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
     defaultPluginData,
     pluginDefaults,
     onComponentMount,
-    initialIntent,
     languageDir,
     locale,
     shouldRenderOptimizedImages,
     siteDomain,
+    sandboxedDomain,
     setInPluginEditingMode,
     getInPluginEditingMode,
     getEditorState,
@@ -86,16 +88,34 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       uiSettings: config.uiSettings,
       getToolbarSettings: config.getToolbarSettings,
       getEditorBounds,
-      initialIntent,
       languageDir,
       locale,
       shouldRenderOptimizedImages,
       siteDomain,
+      sandboxedDomain,
       setInPluginEditingMode,
       getInPluginEditingMode,
       getEditorState,
       setEditorState,
     });
+
+  const externalizedButtonProps = config?.toolbar?.InsertButtons?.map(button =>
+    generateInsertPluginButtonProps({
+      blockType: config.type,
+      button,
+      helpers,
+      pubsub,
+      commonPubsub,
+      settings,
+      t,
+      isMobile,
+      pluginDefaults,
+      getEditorState,
+      setEditorState,
+      hidePopup: helpers?.closeModal,
+      toolbarName: TOOLBARS.EXTERNAL,
+    })
+  );
   const InsertPluginButtons =
     settings.showInsertButtons &&
     config?.toolbar?.InsertButtons?.map(button => ({
@@ -110,11 +130,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
         t,
         isMobile,
         pluginDefaults,
-        initialIntent,
         languageDir,
         locale,
         shouldRenderOptimizedImages,
         siteDomain,
+        sandboxedDomain,
         setInPluginEditingMode,
         getInPluginEditingMode,
         getEditorState,
@@ -143,11 +163,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       getEditorBounds,
       disableRightClick,
       onComponentMount,
-      initialIntent,
       languageDir,
       locale,
       shouldRenderOptimizedImages,
       siteDomain,
+      sandboxedDomain,
       setInPluginEditingMode,
       getInPluginEditingMode,
       getEditorState,
@@ -190,6 +210,7 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
     InlinePluginToolbar,
     Toolbar,
     InsertPluginButtons,
+    externalizedButtonProps,
     InlineModals,
     TextButtonMapper,
     pubsub,