+++ packages/viewer/web/src/getPluginViewers.jsx
@@ -141,6 +141,7 @@ PluginViewer.propTypes = {
     disabled: PropTypes.bool,
     seoMode: PropTypes.bool,
     siteDomain: PropTypes.string,
+    sandboxedDomain: PropTypes.string,
     disableRightClick: PropTypes.bool,
   }).isRequired,
 };