+++ plugins-bundle-analyzer/src/viewer-with-wrapper.js
@@ -1,23 +1,11 @@
 import React from 'react';
-import { RichContentWrapper } from 'wix-rich-content-wrapper';
-import { htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer.js';
-//import { pluginButton } from 'wix-rich-content-plugin-button/dist/module.viewer.js';
-import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer.js';
-//import { pluginDivider } from 'wix-rich-content-plugin-divider/dist/module.viewer.js';
-//import { pluginGallery } from 'wix-rich-content-plugin-gallery/dist/module.viewer.js';
-import { RichContentViewer } from 'wix-rich-content-viewer';
-import { linkTypeMapper } from 'wix-rich-content-plugin-link/dist/module.viewer.js';
+import { RicosViewer } from 'ricos-viewer';
+import { pluginHtml } from 'wix-rich-content-plugin-html/dist/module.viewer.js';
+import { pluginImage } from 'wix-rich-content-plugin-image/dist/module.viewer.js';
+import { pluginLink } from 'wix-rich-content-plugin-link/dist/module.viewer.js';
 
-const typeMappers = [imageTypeMapper, linkTypeMapper, htmlTypeMapper];
+const plugins = [pluginImage, pluginLink, pluginHtml];
 
 export default () => {
-  return (
-    <RichContentWrapper
-      withFullscreen={false}
-      //plugins={[pluginButton(), pluginDivider(), pluginGallery(), pluginHtml(), pluginImage()]}
-      theme={'Default'}
-    >
-      <RichContentViewer typeMappers={typeMappers} />
-    </RichContentWrapper>
-  );
+  return <RicosViewer plugins={plugins} theme={'Default'} />;
 };