+++ CHANGELOG.md
@@ -13,22 +13,45 @@
   <summary>
     Changes that have landed in master but are not yet released.
     Click to see more.
+
   </summary>
 
 
 ### :rocket: New Feature
+- `editor` `viewer`
+  - [#1091](https://github.com/wix-incubator/rich-content/pull/1091) add sandboxedDomain prop for iframe security
 - `editor`
   - [#1041](https://github.com/wix-incubator/rich-content/pull/1041) shift+tab click deletes tab character ('\t')
   - [#1065](https://github.com/wix-incubator/rich-content/pull/1065) backspace click at start of block decrease indentation
-- `fullscreen`
-  - [#845](https://github.com/wix-incubator/rich-content/pull/845) fullscreen image titles and design changes
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) external toolbar API added
+- `wrapper`
+  - [#1003](https://github.com/wix-incubator/rich-content/pull/1003) RichContentWrapper was split into RicosEditor & RicosViewer
+- `viewer`
+  - [#1088](https://github.com/wix-incubator/rich-content/pull/1048) viewerAction callback in helpers is now supported, triggered on image expand (gallery & viewer)  
 ### :bug: Bug Fix
+- `wrapper` `viewer`
+  - [#1101](https://github.com/wix-incubator/rich-content/pull/1101) fix hashtag decorator in viewer
+- `plugins`
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) svg icon id conflicts fixed
+- `giphy`
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) giphy toolbar not showing after giphy is added 
 - `link-preview`
   - [#1048](https://github.com/wix-incubator/rich-content/pull/1048) fix: enable link preview and link embed by default
+- `common`
+  - [#1092](https://github.com/wix-incubator/rich-content/pull/1092) text indentation not showing in rtl
 ### :house: Internal
 - `exampleApp`
   - [0f849222](https://github.com/wix-incubator/rich-content/commit/0f849222) fix translations
-  
+- `storybook`
+  - [#1003](https://github.com/wix-incubator/rich-content/pull/1003) added story for wrapper handling of static text toolbar
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) external plugin sidebar added
+- `storybook`
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) initial intent example added
+- `editor`
+  - [#1078](https://github.com/wix-incubator/rich-content/pull/1078) insert buttons at the beginning of the inline toolbar
+- `editor` `viewer`
+  - [#1090](https://github.com/wix-incubator/rich-content/pull/1090) support normalize config with {disableInlineImages: false/true}
+
 </details>
 <hr/>
 
@@ -49,7 +72,6 @@
   - [#898](https://github.com/wix-incubator/rich-content/pull/898) text indentation
 - `viewer`
   - [#1005](https://github.com/wix-incubator/rich-content/pull/1005) add viewMode SEO to ProGallery
-
 ### :bug: Bug Fix
 - `social-modals`
   - [#1037](https://github.com/wix-incubator/rich-content/pull/1037) disable text input autocomplete